Fix/wait for service (#56)
<!-- Thanks for the contribution, this is awesome. -->

# PR Details
## Description
Current implementation of carma-carla-integration's nodes doesn't wait
for the ROS service to become available and only tries once. When
platform's route node services become available, it is not yet actually
able to perform those services such as get_available_routes(),
set_active_route() etc. The route node itself needs time to process the
routes, so the carma-carla-integration tools needed to continuously try
the services again to reliably perform those tasks Previously it was
just relying on sleep() to align the timings of different tools to
become available.

This PR resolves the issue (mostly) where platform would fail to select
route or engage randomly in some of the runs.
Also added use_sim_time to make sure the nodes in this launch use
simulation time (/localization/current_pose previously was printing
real-time header.stamp epoch for example)

Supporting PR (this enables the logging to be real-time and unbuffered
where the developer can see real-time when the vehicle is about to
engage or if there has been any error. Previously, all the logs would
flush after printing nothing for a while):
https://github.com/usdot-fhwa-stol/carma-carla-integration/pull/56
<!--- Describe your changes in detail -->

## Related GitHub Issue
https://github.com/usdot-fhwa-stol/carma-platform/issues/2222
<!--- This project only accepts pull requests related to open GitHub
issues or Jira Keys -->
<!--- If suggesting a new feature or change, please discuss it in an
issue first -->
<!--- If fixing a bug, there should be an issue describing it with steps
to reproduce -->
<!--- Please DO NOT name partially fixed issues, instead open an issue
specific to this fix -->
<!--- Please link to the issue here: -->

## Related Jira Key

<!-- e.g. CAR-123 -->
CDAR-612
https://usdot-carma.atlassian.net/browse/CDAR-612
## Motivation and Context

<!--- Why is this change required? What problem does it solve? -->

## How Has This Been Tested?
Tested at least 10 times without failing on simulation PC 1
Along the changes with this PR, but this PR should be effective
standalone

<!--- Please describe in detail how you tested your changes. -->
<!--- Include details of your testing environment, and the tests you ran
to -->
<!--- see how your change affects other areas of the code, etc. -->

## Types of changes

<!--- What types of changes does your code introduce? Put an `x` in all
the boxes that apply: -->

- [X] Defect fix (non-breaking change that fixes an issue)
- [ ] New feature (non-breaking change that adds functionality)
- [ ] Breaking change (fix or feature that cause existing functionality
to change)

## Checklist:

<!--- Go over all the following points, and put an `x` in all the boxes
that apply. -->
<!--- If you're unsure about any of these, don't hesitate to ask. We're
here to help! -->

- [X] I have added any new packages to the sonar-scanner.properties file
- [X] My change requires a change to the documentation.
- [X] I have updated the documentation accordingly.
- [X] I have read the
[**CONTRIBUTING**](https://github.com/usdot-fhwa-stol/carma-platform/blob/develop/Contributing.md)
document.
- [X] I have added tests to cover my changes.
- [X] All new and existing tests passed.